reorganize server package so we can pull it in for easier integration testing
