+++ server/src/server/api/routes/spawn.rs
