+++ server/src/server/api/routes/status.rs
