+++ server/src/server/api/routes/stop.rs
