+++ server/src/server/api/routes/stream_log.rs
