+++ server/src/server/auth.rs
